# backend/collector-openapi.yml
openapi: 3.0.0
components:
  schemas:
...
    LocationResponse:
      title: LocationResponse
      description: Represents the physical locations where collection items.
      required:
        - name
      type: object
      properties:
        name:
          description: The name of the location (e.g., Living Room Bookshelf, Storage Box).
          type: string
          minLength: 1
        description:
          description: >-
            A description of the location. Provides users with more information
            about each location.
          type: string
          minLength: 1
        userId:
          type: string
        addedAt:
          type: string
          format: date
        lastModified:
          type: string
          format: date
        attachmentUrl:
          description: >-
            A upload url to attach a file to an item or location depending on
            the request parameter.
          type: string
          minLength: 1
        locationId:
          description: A physical location where the item is stored.
          type: string
          minLength: 1
      additionalProperties: false
...
    ErrorResponse:
      title: ErrorResponse
      description: Represents an error.
      required:
        - statusCode
        - errorCode
        - message
      type: object
      properties:
        statusCode:
          description: The http status code.
          type: number
          minimum: 100
          maximum: 599
        errorCode:
          description: An unique error code.
          type: string
          minLength: 1
        message:
          description: Short error message.
          type: string
          minLength: 1
        details:
          description: Additional details about the error (optional).
          type: string
          minLength: 1
      additionalProperties: false
...
    Access-Control-Allow-Origin:
      type: string
      default: '*'
      example: https://developer.mozilla.org
    Access-Control-Allow-Credentials:
      type: boolean
      default: true
...
    locationId:
      type: string
    c8793222-7ccd-49fc-acf1-8f22eb1cb413:
      title: UpdateLocationRequest
      description: Represents the physical locations where collection items.
      required:
        - name
      type: object
      properties:
        name:
          description: The name of the location (e.g., Living Room Bookshelf, Storage Box).
          type: string
          minLength: 1
        description:
          description: >-
            A description of the location. Provides users with more information
            about each location.
          type: string
          minLength: 1
      additionalProperties: false
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
info:
  title: Collector API
  description: Serverless application for collection management
  version: v.0.0.0
paths:
...
  /locations/{locationId}:
    patch:
      summary: Update location
      description: Updates a location with new informations.
      operationId: UpdateLocation
      parameters:
        - name: locationId
          in: path
          description: ID of the location which will receive the update.
          required: true
          schema:
            $ref: '#/components/schemas/locationId'
      tags:
        - location
      security:
        - bearerAuth: []
      requestBody:
        description: ''
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/c8793222-7ccd-49fc-acf1-8f22eb1cb413'
      responses:
        '200':
          description: The updated location with changed parameters.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocationResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '400':
          description: Bad Request - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '401':
          description: Unauthorized - user is not logged in or authorization failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '404':
          description: Not Found - message with additional information what wasn't found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '500':
          description: Internal Server Error - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
    delete:
      summary: Delete location
      description: Deletes a location of a user.
      operationId: DeleteLocation
      parameters:
        - name: locationId
          in: path
          description: ID of the location which will be deleted.
          required: true
          schema:
            $ref: '#/components/schemas/locationId'
      tags:
        - location
      security:
        - bearerAuth: []
      responses:
        '204':
          description: No Content.
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '400':
          description: Bad Request - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '401':
          description: Unauthorized - user is not logged in or authorization failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '404':
          description: Not Found - message with additional information what wasn't found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
        '500':
          description: Internal Server Error - message with additional information.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          headers:
            Access-Control-Allow-Origin:
              description: >-
                The Access-Control-Allow-Origin response header indicates
                whether the response can be shared with requesting code from the
                given
                [origin](https://developer.mozilla.org/en-US/docs/Glossary/Origin).
                - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Origin)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Origin'
            Access-Control-Allow-Credentials:
              description: >-
                The Access-Control-Allow-Credentials response header tells
                browsers whether to expose the response to the frontend
                JavaScript code when the request's credentials mode
                ([Request.credentials](https://developer.mozilla.org/en-US/docs/Web/API/Request/credentials))
                is include. - [MDN
                Link](https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Access-Control-Allow-Credentials)
              schema:
                $ref: '#/components/schemas/Access-Control-Allow-Credentials'
...